Move KernelData and QuadratureRule to base
